requirements.txt
fastapi
uvicorn[standard]
httpx
loguru
python-dotenv
pydantic
typing_extensions
sqlalchemy[asyncio]
asyncpg
jinja2
